adjust e2e
