+++ packages/ng-helper-vscode/package.json
@@ -115,11 +115,13 @@
     "plugin:compile": "cd ../typescript-plugin && yarn compile",
     "plugin:clean": "node scripts/resolver.js --clean",
     "compile": "yarn tsc && yarn lint && node esbuild.js --production",
+    "compile:e2e": "yarn tsc && yarn lint && node esbuild.js",
     "watch:esbuild": "yarn plugin:clean && node esbuild.js --watch",
     "tsc": "tsc --noEmit",
     "lint": "eslint src --ext ts",
     "test": "yarn test:e2e",
-    "test:e2e": "yarn e2e:clean && yarn e2e:compile && cd tests && node ./dist/runTest.js",
+    "test:e2e": "yarn e2e:prepare && cd tests && node ./dist/runTest.js",
+    "e2e:prepare": "yarn compile:e2e && yarn e2e:clean && yarn e2e:compile",
     "e2e:clean": "cd tests && rm -rf dist && rm -f .vscode-test/user-data/1.93-main",
     "e2e:compile": "cd tests && tsc"
   },
