+++ packages/ng-helper-vscode/tests/testUtils.ts
@@ -1,26 +1,37 @@
 import * as vscode from 'vscode';
 
-import { MY_EXTENSION_ID } from './testConstants';
+import { BAR_FOO_COMPONENT_HTML_PATH, BAR_FOO_COMPONENT_TS_PATH, COMPLETION_COMMAND } from './testConstants';
 
 export function sleep(ms: number): Promise<void> {
     return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-export async function confirmExtensionActive() {
-    const extension = vscode.extensions.getExtension(MY_EXTENSION_ID);
-    if (!extension) {
-        console.log('====> Extension found!!!');
-        return;
+export async function activate() {
+    // enable tsserver
+    await vscode.window.showTextDocument(vscode.Uri.file(BAR_FOO_COMPONENT_TS_PATH), { preview: false });
+    await sleep(1000);
+
+    // show the document
+    await vscode.window.showTextDocument(vscode.Uri.file(BAR_FOO_COMPONENT_HTML_PATH));
+    await waitForCompletionToBeAvailable(20);
+}
+
+async function waitForCompletionToBeAvailable(maxSeconds: number) {
+    let tries = 0;
+    while (tries < maxSeconds) {
+        const position = new vscode.Position(0, 9);
+        // For a complete list of standard commands, see
+        // https://code.visualstudio.com/api/references/commands
+        const completion = await vscode.commands.executeCommand<vscode.CompletionList>(
+            COMPLETION_COMMAND,
+            vscode.Uri.file(BAR_FOO_COMPONENT_HTML_PATH),
+            position,
+            ' ', // space to trigger completion
+        );
+        if (completion && completion.items.length > 0) {
+            return;
+        }
+        tries++;
+        await sleep(1000);
     }
-
-    console.log('Extension ID:', extension.id);
-
-    if (extension && !extension.isActive) {
-        // 如果扩展未激活，等待其激活
-        await extension.activate();
-        // 给更多时间让 VS Code 和扩展完全初始化
-        await sleep(2_000);
-    }
-
-    console.log('Is extension active:', extension.isActive);
 }
